Syntax error

+++ Heaviside/Heaviside/op_host/heaviside.cpp
@@ -13,7 +13,7 @@ static ge::graphStatus TilingFunc(gert::TilingContext* context)
     HeavisideTilingData tiling;
 
     auto ascendcPlatform = platform_ascendc::PlatformAscendC(context->GetPlatformInfo());
-    uint64_t size_of_dtype;
+    uint32_t size_of_dtype;
     ge::TypeUtils::GetDataTypeLength(context->GetInputDesc(0)->GetDataType(), size_of_dtype);
     auto num_cores = ascendcPlatform.GetCoreNumAiv();
     context->SetBlockDim(num_cores);
